+++ src/NodeTrait.php
@@ -11,7 +11,8 @@ namespace DecodeLabs\Atlas;
 
 use DateInterval;
 use DateTime;
-use DecodeLabs\Exceptional;
+use DecodeLabs\Coercion;
+use Stringable;
 
 trait NodeTrait
 {
@@ -35,8 +36,9 @@ trait NodeTrait
     /**
      * Normalize dots in a path
      */
-    protected function normalizePath(string $path): string
-    {
+    protected function normalizePath(
+        string $path
+    ): string {
         $root = $path[0] === '/' ? '/' : '';
         $parts = explode('/', trim($path, '/'));
         $output = [];
@@ -59,8 +61,9 @@ trait NodeTrait
     /**
      * Compare last modified
      */
-    public function hasChanged(int $seconds = 30): bool
-    {
+    public function hasChanged(
+        int $seconds = 30
+    ): bool {
         if (!$this->exists()) {
             return false;
         }
@@ -71,18 +74,15 @@ trait NodeTrait
     /**
      * Compare with interval string
      */
-    public function hasChangedIn(string $timeout): bool
-    {
-        if (preg_match('/^[0-9]+$/', $timeout)) {
+    public function hasChangedIn(
+        DateInterval|string|Stringable|int $timeout
+    ): bool {
+        if (is_int($timeout)) {
             return $this->hasChanged((int)$timeout);
         }
 
         $date = new DateTime('now');
-
-        if (!$interval = DateInterval::createFromDateString($timeout)) {
-            throw Exceptional::InvalidArgument('Invalid interval string: ' . $timeout);
-        }
-
+        $interval = Coercion::toDateInterval($timeout);
         $ts = $date->sub($interval)->getTimestamp();
 
         if (!$mod = $this->getLastModified()) {